requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
Django==5.2.6
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
openai==2.0.0
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==25.1
google-generativeai==0.8.3
python-dotenv==1.0.0
langchain==0.3.15
langchain-openai==0.2.14
langchain-core==0.3.28
